Reject requests, handling of multiple requests
Updated freemarker to 2.3.28 from 2.3.23 on pm.xml to be able to use maps in ftls

When a request is acceped by someone else, render an error indicating that the request is no longer valid.

Fixed the issue with request to one user showing to everyone. Using maps in gamelobby.ftl